Move types/structs/enums to `types.rs`
Extract all pallet-level type definitions from `lib.rs` into a new
`types.rs` module. No behavior change.

Moved (and re-exported via `pub use types::*;`):
- `CreditOf`, `AuthorizationExtent` (+ `has_permanent_capacity`),
  `AuthorizationScope`, `AuthorizedCaller`, `AuthorizedCallerFor`,
  `TransactionKind`, `TransactionInfo` (+ `total_chunks`).

Moved with widened `pub(crate)` visibility (were private):
- `BalanceOf`, `AuthorizationScopeFor`, `Authorization` (+ fields +
  `expired`), `AuthorizationFor`, `CheckContext` (+ methods), and the
  previously-private `TransactionInfo::chunk_root` and
  `TransactionInfo::block_chunks` fields.
